middleware

+++ app/Http/Controllers/PostsController.php
@@ -7,7 +7,12 @@ use App\Post;
 use DB;
 
 class PostsController extends Controller
-{
+{    
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -15,7 +20,7 @@ class PostsController extends Controller
      */
     public function index()
     {
-        $posts = Post::orderBy('created_at', 'desc')->paginate(2);
+        $posts = Post::orderBy('created_at', 'desc')->paginate(3);
         return view('posts.index')->with('posts', $posts);
     }
 
@@ -70,6 +75,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -102,6 +110,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed');
     }
